Do not modify LD_LIBRARY_PATH when setting install_path
Previously, it was required to add <root>/lib to LD_LIBRARY_PATH because
CADET-Core was not setting the RPATH correctly. This was fixed in CADET
v4.4.0
Since the lib path was simply appended to os.environ['LD_LIBRARY_PATH']
this could lead to inconsistent behaviour when trying to compare
different versions of CADET-Core and other versions were not properly
removed.